+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/ExoPlayerImplInternal.java
@@ -185,6 +185,7 @@ import java.util.Objects;
   private static final int MSG_SEEK_COMPLETED_IN_SCRUBBING_MODE = 37;
   private static final int MSG_SET_SCRUBBING_MODE_PARAMETERS = 38;
   private static final int MSG_SET_IMAGE_METADATA_LISTENER = 39;
+  private static final int MSG_SET_AUDIO_SESSION_ID = 40;
 
   private static final long BUFFERING_MAXIMUM_INTERVAL_MS =
       Util.usToMs(Renderer.DEFAULT_DURATION_TO_PROGRESS_US);
@@ -544,6 +545,10 @@ import java.util.Objects;
     handler.obtainMessage(MSG_SET_VOLUME, volume).sendToTarget();
   }
 
+  public void setAudioSessionId(int audioSessionId) {
+    handler.obtainMessage(MSG_SET_AUDIO_SESSION_ID, audioSessionId, 0).sendToTarget();
+  }
+
   private void handleAudioFocusPlayerCommandInternal(
       @AudioFocusManager.PlayerCommand int playerCommand) throws ExoPlaybackException {
     updatePlayWhenReadyWithAudioFocus(
@@ -837,6 +842,9 @@ import java.util.Objects;
         case MSG_SET_VOLUME:
           setVolumeInternal((Float) msg.obj);
           break;
+        case MSG_SET_AUDIO_SESSION_ID:
+          setAudioSessionIdInternal(msg.arg1);
+          break;
         case MSG_AUDIO_FOCUS_PLAYER_COMMAND:
           handleAudioFocusPlayerCommandInternal(/* playerCommand= */ msg.arg1);
           break;
@@ -1110,6 +1118,12 @@ import java.util.Objects;
     float scaledVolume = volume * audioFocusManager.getVolumeMultiplier();
     for (RendererHolder renderer : renderers) {
       renderer.setVolume(scaledVolume);
+    }
+  }
+
+  private void setAudioSessionIdInternal(int audioSessionId) throws ExoPlaybackException {
+    for (RendererHolder renderer : renderers) {
+      renderer.setAudioSessionId(audioSessionId);
     }
   }